create db_loader, make dg use it, it should run now
