+++ dadguide/database_manager.py
@@ -18,12 +18,6 @@ import tsutils
 
 logger = logging.getLogger('red.padbot-cogs.dadguide.database_manager')
 
-# def _data_file(file_name: str) -> str:
-#     return os.path.join(str(data_manager.cog_data_path(raw_name='dadguide')), file_name)
-#
-# DB_DUMP_FILE = _data_file('dadguide.sqlite')
-# DB_DUMP_WORKING_FILE = _data_file('dadguide_working.sqlite')
-
 
 class Attribute(Enum):
     """Standard 5 PAD colors in enum form. Values correspond to DadGuide values."""
@@ -79,17 +73,6 @@ class Server(Enum):
 class DadguideTableNotFound(Exception):
     def __init__(self, table_name):
         self.message = '{} not found'.format(table_name)
-
-
-def load_database(existing_db):
-    # Release the handle to the database file if it has one
-    if existing_db:
-        existing_db.close()
-    # Overwrite the working copy so we can open a handle to it without affecting future downloads
-    if os.path.exists(DB_DUMP_FILE):
-        shutil.copy2(DB_DUMP_FILE, DB_DUMP_WORKING_FILE)
-    # Open the new working copy.
-    return DadguideDatabase(data_file=DB_DUMP_WORKING_FILE)
 
 
 class DadguideDatabase(object):
